Some SessionHandler cleanup
